Adds correct click handler for getting one page
Dropdown list shows text cursor when hovering, but clicking calls
onGetOnePage in events.js. No need for extra function outside of the
pre-existing document ready function. Adds console.logs to confirm
onGetOnePage is being called. Data from getFormFields returns an object,
not sure if it can be used to get the data-id to make function operate
the way it's intended.